Move Replicated*** to replicated_storage
